+++ uima-ducc-transport/src/main/java/org/apache/uima/ducc/transport/event/common/IDuccReservation.java
@@ -36,8 +36,8 @@ public interface IDuccReservation extends Serializable {
 	public NodeIdentity getNodeIdentity();
 	public void setNodeIdentity(NodeIdentity nodeIdentity);
 	
-	public int getBytes();
-	public void setBytes(int bytes);
+	public long getBytes();
+	public void setBytes(long bytes);
 	
 	public ITimeWindow getTimeWindow();
 	public void setTimeWindow(ITimeWindow timeWindow);
